Changed context depth to 3
Hi again - when I used `String.join()` the other day in my commit, it was like 3am in my timezone so I was very sleepy and didn't realize that my pull request was going into my own fork. I needed to click another button afterward to forward the pull request to your actual repo here. Since adding the extra depth will add even more context to the linked posts and since I'd at least like the credit of having the badge of honor of a pull request to hang on my fireplace I thought I'd propose this file change. 

A lot of the deltas that are posted from what I can tell are the result of longer comment chains anyway and it's what the /r/bestof and /r/DepthHub sbureddits recommend in their rules, so I figure that it'd be sufficient here as well.